Ultimate puzzle solving palooza

+++ mge_v18_student_version/src/main.cpp
@@ -20,7 +20,7 @@
  */
 int main()
 {
-	std::srand(time(NULL));
+	std::srand((unsigned int)time(NULL));
     std::cout << "Starting Game" << std::endl;
 
     AbstractGame* game = new TPerson();
